// src/app/types/buttons.ts
export type LinkButtonSettings = {
    type: 'link';
    /** Label displayed inside the button */
    text: string;
    /** Button decoration (icon, arrow) or left blank for none */
    decoration?: BtnDecoration;
    /** React Router path or absolute URL the button links to */
    link: string;
    /** Whether the link opens in a new tab @default '_self' */
    target?: "_blank" | "_self";
}

export type ModalButtonSettings = {
    type: 'modal';
    /** Label displayed inside the button */
    text: string;
    /** Button decoration (icon, arrow) or left blank for none */
    decoration?: BtnDecoration;
    /** */
    modalID: string;
    /** */
    modalContent: Omit<ModalProps, 'isOpen' | 'onClose'>;
}

/**
 * Button for visual only (not clickable)
 * 
 */
export type VisualButtonSettings = {
    type: 'visual';
    /** Label displayed inside the button */
    text: string;
    /** Button decoration (icon, arrow) or left blank for none */
    decoration?: BtnDecoration;
}

export type ButtonSettingProps = LinkButtonSettings | ModalButtonSettings | VisualButtonSettings;
// ...
export type ButtonProps = WithHTMLProps &
    WithA11yProps & {
        /** Text content and link destination */
        btnSettings: ButtonSettingProps;

        /** Visual style variant @default 'solid' */
        variant?: BtnVariants;
        /** Color theme applied to the button and arrow */
        colorScheme?: BtnColorScheme;
        /** Adds optional full-width class */
        fullWidth?: boolean;
    };


// export type ButtonWithDecorationProps = {

// }

// #region Button Variants and Variants Map

/**
 *
 */
export type BtnVariants = "solid" | "outline" | "lines";

/**
 *
 * @template N - The required number of elements
 */
export type BtnVariantMap<N extends number> = RequireX<BtnVariants, N>;

// #endregion

// #region Button Color Scheme and Theme Map

/**
 *
 */
export type BtnColorScheme = "gold" | "cream" | "black";